fix file hook
