+++ src/pages/homePage/componentsHome/hooks/useProvincesQuery.ts
@@ -1,6 +1,12 @@
-import { findProvinces } from './../InjectionSite';
+import { Province } from './../InjectionSite';
 import { useQuery } from '@tanstack/react-query';
 import { QueryKey } from '../../../../hooks/QueryKey';
+import api from '../../../../utils/axios/instance';
+
+export const findProvinces = async () => {
+  const res = await api.get<Province[]>('/administrative-unit/provinces');
+  return res.data;
+};
 
 export const useProvincesQuery = () => {
   return useQuery({
